+++ lab5/LiaoWC/validate.sh
